Fixed the following issues:- 1. Removed the default value check from vendor vector encoder. 2. Added a non-empty check for vector in OOBVendorEncoder class.

+++ src/main/java/com/iabtcf/gdpr/phase2/Utils/TCModel.java
@@ -247,7 +247,7 @@ public class TCModel {
             //throw error
         }
         gvlMap.forEach((key,value)->{
-            vector.getSet().add((Integer) key);
+            vector.getSet().add(key);
         });
         vector.setBitLength(0);
     }
